resume-ml/requirements.txt
Flask==3.0.0
transformers==4.36.0
torch==2.5.1
sentence-transformers==2.3.1
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn==1.4.0
requests-toolbelt==1.0.0
